+++ components/interview/components/Prompts/Prompts.tsx
